requirements.txt
GitPython
langchain
sentence_transformers
python-dotenv
chromadb
flask
ctransformers